Extraer Testimonios Arreglado error en react-slick Extraido testimonios en su componente Importado componente en page
